file attribute changes, about app, and email app overhaul

+++ buildImages.sh
@@ -1,3 +1,9 @@
+#!/bin/bash
+
+if [[ $(id -u) == 0 ]]; then
+  echo "SUDO will bop the owner on all the asset files."
+  exit
+fi
 rm ./Assets/icons/*.png
 cp ./AssetMasters/*.png ./Assets/icons/
 python3 ImageSplit.py
